+++ client/src/component/molecule/index.ts
@@ -1,5 +1,6 @@
 import ButtonDiv from './ButtonDiv'
 import HeaderInput from './HeaderInput'
 import Modal from './Modal'
+import ActionMenuButton from './ActionMenuButton'
 
-export default { ButtonDiv, HeaderInput, Modal }
+export default { ButtonDiv, HeaderInput, Modal, ActionMenuButton }
