+++ client/src/component/organism/ActionBar/ActionBar.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import ActionBar from '.'
 
 export default {
-  title: 'Molecule/ActioinBar',
+  title: 'Organism/ActionBar',
   component: ActionBar,
 }
 
@@ -12,6 +12,14 @@ export const actionBar = () => {
   const targetId = 1
   const targetAuthorId = 1
   const loginUserId = 1
+
+  const handleDeleteButtonClick = () => {
+    alert(`Delete message [messageId: ${targetId}]`)
+  }
+  const handleEditButtonClick = () => {
+    alert(`Edit message [messageId: ${targetId}]`)
+  }
+
   return (
     <>
       <div>targetAuthorId === loginUserId</div>
@@ -20,6 +28,8 @@ export const actionBar = () => {
         targetId={targetId}
         targetAuthorId={targetAuthorId}
         loginUserId={loginUserId}
+        onDeleteButtonClick={handleDeleteButtonClick}
+        onEditButtonClick={handleEditButtonClick}
       />
       <hr />
       <div>targetAuthorId === loginUserId & targetType is MESSAGE </div>
@@ -28,6 +38,8 @@ export const actionBar = () => {
         targetId={targetId}
         targetAuthorId={targetAuthorId}
         loginUserId={loginUserId}
+        onDeleteButtonClick={handleDeleteButtonClick}
+        onEditButtonClick={handleEditButtonClick}
       />
       <hr />
       <div>targetAuthorId !== loginUserId</div>
@@ -36,6 +48,8 @@ export const actionBar = () => {
         targetId={targetId}
         targetAuthorId={targetAuthorId + 1}
         loginUserId={loginUserId}
+        onDeleteButtonClick={handleDeleteButtonClick}
+        onEditButtonClick={handleEditButtonClick}
       />
       <hr />
       <div>targetAuthorId !== loginUserId & targetType is MESSAGE </div>
@@ -44,6 +58,8 @@ export const actionBar = () => {
         targetId={targetId}
         targetAuthorId={targetAuthorId + 1}
         loginUserId={loginUserId}
+        onDeleteButtonClick={handleDeleteButtonClick}
+        onEditButtonClick={handleEditButtonClick}
       />
     </>
   )
