+++ client/src/component/organism/index.ts
@@ -1,4 +1,5 @@
 import MessageCard from './MessageCard'
 import Header from './Header'
+import MessageActionsMenu from './MessageActionsMenu'
 
-export default { MessageCard, Header }
+export default { MessageCard, Header, MessageActionsMenu }
